Add files via upload
Here is a t-test project, t-test is used to compare the means of two groups to determine if a statistically significant difference exists between them. It's a type of hypothesis testing that quantifies how significant the difference between the means is, taking into account the variance within the groups. There are different types of t-tests, such as the independent-samples t-test, paired-samples t-test for the same group measured at different times. Here i Completed 3 Types of t test. firstly perform one sample t test to see the two variables average difference whether 40 or not, and the result will show the average value is not 40. this test i use titanic data sets